rewrite P4
