+++ abramov.vladislav/P4/str_functions.hpp
@@ -3,7 +3,6 @@
 #include <iostream>
 namespace abramov
 {
-  char *getStr(std::istream &in);
   char *unionStrs(char *destination, const char *source);
 }
 #endif
