+++ abramov.vladislav/common/getString.hpp
@@ -1,0 +1,8 @@
+#ifndef GETSTRING_HPP
+#define GETSTRING_HPP
+#include <iostream>
+namespace abramov
+{
+  char *getStr(std::istream &in);
+}
+#endif
